fix panel bug when download excel
